requirements.txt
datasets
langchain
langchain-community
Markdown
tokenizers
transformers
unsloth==2025.3.6
unsloth_zoo==2025.3.4
unstructured
vllm
wandb
pandas
pymongo
python-dotenv
tqdm
tabulate